% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST21_2022.R
\name{deploy_MAST21_2022}
\alias{deploy_MAST21_2022}
\title{Run the MAST21 2022 protocol}
\usage{
deploy_MAST21_2022(musicassessr_state = "test")
}
\arguments{
\item{musicassessr_state}{}
}
\value{

}
\description{
Run the MAST21 2022 protocol
}
